The authentication schema migration wasn't the right one we were testing with.

+++ authentication/db/migrate/20100913234705_create_refinerycms_authentication_schema.rb
@@ -4,40 +4,45 @@ class CreateRefinerycmsAuthenticationSchema < ActiveRecord::Migration
     create_table ::Refinery::RolesUsers.table_name, :id => false, :force => true do |t|
       t.integer "user_id"
       t.integer "role_id"
-    end
+    end unless ::Refinery::RolesUsers.table_exists?
 
     create_table ::Refinery::Role.table_name, :force => true do |t|
       t.string "title"
+    end unless ::Refinery::Role.table_exists?
+
+    unless ::Refinery::UserPlugin.table_exists?
+      create_table ::Refinery::UserPlugin.table_name, :force => true do |t|
+        t.integer "user_id"
+        t.string  "name"
+        t.integer "position"
+      end
+
+      add_index ::Refinery::UserPlugin.table_name, ["name"], :name => "index_#{::Refinery::UserPlugin.table_name}_on_title"
+      add_index ::Refinery::UserPlugin.table_name, ["user_id", "name"], :name => "index_unique_#{::Refinery::UserPlugin.table_name}", :unique => true
+
     end
 
-    create_table ::UserPlugin.table_name, :force => true do |t|
-      t.integer "user_id"
-      t.string  "name"
-      t.integer "position"
+    unless ::Refinery::User.table_exists?
+      create_table ::Refinery::User.table_name, :force => true do |t|
+        t.string   "login",             :null => false
+        t.string   "email",             :null => false
+        t.string   "crypted_password",  :null => false
+        t.string   "password_salt",     :null => false
+        t.string   "persistence_token"
+        t.datetime "created_at"
+        t.datetime "updated_at"
+        t.string   "perishable_token"
+      end
+
+      add_index ::Refinery::User.table_name, ["id"], :name => "index_#{::Refinery::User.table_name}_on_id"
     end
-
-    add_index ::UserPlugin.table_name, ["name"], :name => "index_#{::UserPlugin.table_name}_on_title"
-    add_index ::UserPlugin.table_name, ["user_id", "name"], :name => "index_unique_#{::UserPlugin.table_name}", :unique => true
-
-    create_table ::User.table_name, :force => true do |t|
-      t.string   "login",             :null => false
-      t.string   "email",             :null => false
-      t.string   "crypted_password",  :null => false
-      t.string   "password_salt",     :null => false
-      t.string   "persistence_token"
-      t.datetime "created_at"
-      t.datetime "updated_at"
-      t.string   "perishable_token"
-    end
-
-    add_index ::User.table_name, ["id"], :name => "index_#{::User.table_name}_on_id"
   end
 
   def self.down
     [::User].reject{|m|
       !(defined?(m) and m.respond_to?(:table_name))
     }.each do |model|
-      drop_table model.table_name
+      drop_table model.table_name if model.table_exists? if model.table_exists?
     end
   end
 end
